refactor droppable gui tests
